refactor(cli): move package manager detection from testing to utils
- Move package manager detection utilities from `src/testing/` to `src/utils/`
- Update CLI utils exports to include package manager utilities
- Remove re-exports from testing module to eliminate duplication
- Update create-trailhead-cli to import from new utils location
- Maintain all functionality while improving architectural clarity

Resolves architectural inconsistency where production code depended on
testing utilities. Package manager detection is now properly located
in utils module as a general utility.

Fixes #49
