+++ CLAUDE.md
@@ -616,7 +616,7 @@ pnpm build                  # Production build
 - `user-first` = Refocus on actual user needs. What do users actually want to accomplish? What's the simplest way to achieve that? Ignore "best practices" that don't serve real user goals.
 - `scope-check` = Verify single responsibility. Does this issue/PR address exactly one problem? Are we trying to do too much? Should this be split into separate issues?
 - `branch-audit` = Check branch structure. Are we working from main? Is this a branch-on-branch (red flag)? Does branch name match issue scope?
-- `gh-issue` = Read GitHub issue and create action plan: 1) Fetch issue details with `gh issue view [number]`, 2) Analyze problem statement and acceptance criteria, 3) Create TodoWrite plan with specific implementation steps, 4) Check for related issues/PRs, 5) Identify affected packages in monorepo, 6) Create feature branch from main with descriptive name.
+- `gh-issue` = Comprehensive GitHub issue analysis and implementation planning: 1) Fetch issue details with `gh issue view [number] --comments` to get full conversation history, 2) Analyze ALL comments for additional context, requirements changes, design decisions, and implementation hints, 3) Deep codebase examination using search tools to identify: existing partial implementations, related code patterns, similar features, affected dependencies, and architectural constraints, 4) Apply core principles analysis: functional patterns, type safety, DRY violations to avoid, YAGNI considerations, and high-ROI testing approach, 5) Cross-reference with related issues/PRs using `gh issue list` and `gh pr list` for context, 6) Identify monorepo impact across all packages and apps, 7) Create comprehensive TodoWrite plan with: problem analysis summary, codebase findings, implementation strategy aligned with principles, specific atomic tasks, testing approach, and affected package considerations, 8) Create feature branch from main with semantic naming (feat/fix/refactor-issue-number-brief-description).
 
 ### Documentation Shortcuts
