+++ packages/cli/src/testing/index.ts
@@ -26,16 +26,3 @@ export {
   testPaths,
   isWindows,
 } from './path-utils.js';
-
-// Package manager utilities
-export {
-  detectPackageManager,
-  getRunCommand,
-  execPackageManagerCommand,
-  getPackageManagerInfo,
-  clearPackageManagerCache,
-  createPackageManagerCache,
-  SemVer,
-  type PackageManager,
-  type DetectOptions,
-} from './package-manager.js';
