+++ packages/cli/src/utils/package-manager.ts
